bz/requirements.txt
appdirs==1.4.3
awscli==1.15.67
awsebcli==3.14.3
beautifulsoup4==4.6.3
bokeh==0.13.0
bs4==0.0.1
cement==2.8.2
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.3.9
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
EasyProcess==0.2.3
Flask==1.0.2
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.6
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
logzero==1.5.0
lyricwikia==0.1.9
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mutagen==1.41.1
mysql-connector==2.1.6
networkx==2.2
numpy==1.15.1
packaging==17.1
pafy==0.5.4
pandas==0.23.4
pathlib==1.0.1
pathspec==0.5.5
Pillow==5.3.0
pipenv==2018.7.1
pyasn1==0.4.4
pyblizzard==1.1
pylint==1.9.1
pyparsing==2.2.0
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
PyVirtualDisplay==0.2.1
PyWavelets==1.0.1
PyYAML==3.13
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
scikit-image==0.14.1
scipy==1.1.0
semantic-version==2.5.0
six==1.11.0
spotdl==1.0.0
spotipy==2.4.4
tabulate==0.7.5
termcolor==1.1.0
titlecase==0.12.0
toolz==0.9.0
tornado==5.1
unicode-slugify==0.1.3
Unidecode==1.0.22
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.10.11
XlsxWriter==1.1.0
youtube-dl==2018.9.10
